+++ lib/notion_databases.dart
@@ -1,5 +1,8 @@
+import 'dart:convert';
+
 import 'package:http/http.dart' as http;
 import 'package:notion_api/base_client.dart';
+import 'package:notion_api/notion/objects/database.dart';
 
 import 'responses/notion_response.dart';
 import 'statics.dart';
@@ -20,6 +23,8 @@ class NotionDatabasesClient extends BaseClient {
   }) : super(token: token, version: version, dateVersion: dateVersion);
 
   /// Retrieve the database with [id].
+  ///
+  /// _See more at https://developers.notion.com/reference/get-database_
   Future<NotionResponse> fetch(String id) async {
     http.Response res =
         await http.get(Uri.https(host, '/$v/$path/$id'), headers: {
@@ -34,6 +39,8 @@ class NotionDatabasesClient extends BaseClient {
   ///
   /// A [startCursor] can be defined to specify the page where to start.
   /// Also a [pageSize] can be defined to limit the result. The max value is 100.
+  ///
+  /// _See more at https://developers.notion.com/reference/get-databases_
   Future<NotionResponse> fetchAll({String? startCursor, int? pageSize}) async {
     Map<String, dynamic> query = {};
     if (startCursor != null) {
@@ -48,6 +55,23 @@ class NotionDatabasesClient extends BaseClient {
       'Authorization': 'Bearer $token',
       'Notion-Version': dateVersion,
     });
+
+    return NotionResponse.fromResponse(res);
+  }
+
+  /// Create a database.
+  ///
+  /// _See more at https://developers.notion.com/reference/create-a-database_
+  Future<NotionResponse> create(Database database) async {
+    http.Response res = await http.post(
+      Uri.https(host, '/$v/$path'),
+      body: jsonEncode(database.toJson()),
+      headers: {
+        'Authorization': 'Bearer $token',
+        'Notion-Version': dateVersion,
+        'Content-Type': 'application/json',
+      },
+    );
 
     return NotionResponse.fromResponse(res);
   }